pyproject.toml
[project]
name = "tw3-chatbot-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "duckduckgo-search>=8.1.1",
    "httpx>=0.28.1",
    "langchain>=0.3.26",
    "langchain-openai>=0.3.27",
    "litellm>=1.74.2",
    "markdownify>=1.1.0",
    "python-dotenv>=1.0.1",
    "mcp[cli]>=1.11.0",
    "fastapi>=0.116.1",
]

[tool.uv.workspace]
members = ["bonus"]
